Data Processing and predictive model scripts/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
Cantera==2.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
dill==0.3.6
dm-tree==0.1.8
etils==0.9.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.7.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.8.2
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
termcolor==2.2.0
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
